requirements.txt
flask
flask_cors
pyjwt
firebase-admin